Replace std::cosf with std::cos

+++ src/gui_editor/menus.cpp
@@ -143,7 +143,7 @@ private:
             using namespace std::chrono;
             auto time =
                     duration_cast<milliseconds>(system_clock::now() - start_t);
-            float wave = std::cosf(time.count() / 120.f);
+            float wave = std::cos(time.count() / 120.f);
             ImVec4 col = ImVec4(0, 0, 1.f, 1.f) +
                          (ImVec4(1.f, 1.f, 0, 0) * (0.2f * wave + 0.8f));
             ImGui::PushStyleColor(ImGuiCol_Text, col);
